2010-11-06  Maurits Rijk  <maurits.rijk@gmail.com>
	* lib/ParamDefList.cs
	* lib/ParamDef.cs
	* lib/Tattoo.cs
	* lib/CoordinateList.cs: minor clean-up

	* lib/Channel.cs: replaced _ID by ID

	* lib/Buffer.cs: minor clean-up

	* lib/Gradient.cs
	* lib/Palette.cs: changed _name to Name

	* lib/DataObject.cs: made _name private instead of protected

	* plug-ins/UnitTest/TestBrush.cs: use BrushInfo class

	* lib/Brush.cs: use BrushInfo class instead of bunch of out
	variables in GetInfo

	* lib/BrushInfo.cs: new file

	* lib/Makefile.am: changed accordingly

	* lib/Brush.cs
	* lib/BoolMatrix.cs
	* plug-ins/Raindrops/RaindropFactory.cs
	* lib/ProcedureSet.cs: clean-up

+++ plug-ins/Forge/Planet.cs
@@ -361,7 +361,7 @@ namespace Gimp.Forge
     RGB RenderWater(double val)
     {
       byte r = (byte)((val > 64) ? (val - 64) * 4 : 0);
-      return new RGB(r, r, 255);
+      return new RGB(r, r, (byte) 255);
     }
 
     void RenderPolarIceCaps(double val, double icet, RGB rgb)
